+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * :fire: Windowx getDisplayRotation, getStatusBarHeight, hasNavigationBar, getNavigationBarHeight, getNavigationBarWidth method moved to Displayx
 * :sparkles: Displayx add getRotation
 * :sparkles: InputMethodx add delayShowSoftInput method
+* :fire: Androidx remove isAtLeastI, isAtLeast14, isAtLeast4_0, isAtLeastIMR1, isAtLeast15, isAtLeast4_0_3 method, getSdkVersionName rename to getVersionName